--- server/models/index.js
@@ -1,9 +1,0 @@
-const User = require('./user');
-const Guild = require('./guild');
-const PromptError = require('./promptError');
-
-module.exports = {
-    User,
-    Guild,
-    PromptError
-};
