--- server/routes/index.js
@@ -1,6 +1,0 @@
-const router = require('express').Router();
-const api = require('./api')
-
-router.use('/api', api)
-
-module.exports = router;
