+++ src/config/mongo/index.ts
@@ -1,5 +1,6 @@
-const mongoose = require('mongoose');
+import mongoose from 'mongoose';
 
 mongoose.connect(process.env.MONGODB_URI || 'mongodb://127.0.0.1:27017/bully-me-gpt');
 
-module.exports = mongoose.connection;
+export default mongoose.connection;
+
