+++ tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "compilerOptions": {
+      "module": "NodeNext",
+      "moduleResolution": "NodeNext",
+      "target": "ES2020",
+      "sourceMap": true,
+      "outDir": "dist",
+    },
+    "include": ["src/**/*"],
+}
